Split auto test palettePropagation2()
palettePropagation was a fairly long test function and was testing more
than one thing at once.

Split palettePropagation2() in two, creating the new auto test
palettePropagation3() that verify if the text color propagation is
correctly being ignored when setting a new palette for buttons. This was
previously tested in palettePropagation2().

palettePropagation2() now only tests if colors are propagated correctly.

Both function are now shorter, easier to understand and maintain.

Change-Id: Icdad272d4cdc437595c0a31ab231d134cb3b5d86
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>