Quick for Android: Call quitQt() after main() exits
In Quick for Android case, if Qt exits itself (Qt.exit() in QML, etc),
we get deadlock in startQtApplication after the user-provided main()
exits.

This is due to terminateQt waiting for a semaphore that - in a normal
Q4A context - is called when the hosting Activity itself is destroyed
from another thread via an activity status callback.

Task-number: QTBUG-130342
Change-Id: I9cd701e0ea86a445e13a6568c4954de6e356e98a
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>