QQ4A: All Loaders extract metadata and applicationArguments
Just like regular Q4A apps and Android service apps, we should also
extract and set the common environment variables and app parameters for
QQ4A apps, so call extractContextMetadata() from the QtEmbeddedLoader
ctor.

In order for QtTest library to know where to output test results,
androidtestrunner passes it some parameters via "applicationArguments"
when launching the application. These params should be passed from the
launch intent to the main application, so now the base QtLoader class
parses this info in extractContextMetadata(), so that it's not just
parsed by QtActivityLoader.

Change-Id: If0444847676d197689b087c5f6684552f8ea73aa
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>